pages-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-multipart==0.0.20
orjson==3.10.13
tenacity==9.0.0
cachetools==5.5.0
structlog==25.1.0
mutagen==1.47.0
ebooklib==0.18
beautifulsoup4==4.12.3
pillow==11.1.0
pypdf==5.1.0
rarfile==4.2
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
rapidfuzz==3.10.1
